requirements.txt
fastapi
uvicorn
openai-whisper
torch
gtts
python-dotenv
ffmpeg-python
pydub
requests
